worker/pyproject.toml
[project]
name = "politicoresto-weighting"
version = "0.1.0"
description = "Survey reweighting worker for PoliticoResto polls — Deville-Särndal calibration via samplics."
requires-python = ">=3.12,<3.13"

dependencies = [
  # Peer-reviewed calibration library (JOSS 2021, validated against R `survey`).
  "samplics>=0.6.0,<0.7",
  # Numerics.
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.12",
  # Operational.
  "structlog>=24.1",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "pynsee>=0.2.5",
]

[dependency-groups]
dev = [
  "pytest>=8.1",
  "pytest-cov>=5.0",
  "pytest-regressions>=2.5",
  "hypothesis>=6.100",
  "ruff>=0.4",
  "mypy>=1.10",
  "pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Safe, explicit defaults. Keep narrow until we need more.
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RET", "N"]
ignore = [
  # Allow ambiguous unicode (we use INSEE labels with accents).
  "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501", "B018", "PLR2004"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
# samplics has partial type stubs; we pin our wrapper's public API typed and
# shield internal library calls with explicit casts where necessary.
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "--strict-markers",
  "--strict-config",
  "-ra",
  # Coverage reporting lives in CI; keep local runs fast.
]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
  "property: Hypothesis-based property tests (may be slower).",
  "differential: Wrapper-vs-raw-samplics parity tests.",
  "golden: Frozen-output regression tests via pytest-regressions.",
  "external_benchmark: Cross-checks against R `survey` or other published outputs.",
  "stress: Random-panel R-vs-Python parity. Requires R installed. Slower (~30s/50 scenarios).",
  "integration: End-to-end tests against local Supabase (requires `supabase start`).",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weighting"]
